LayoutTests/webgpu/webgpu-basics.html

<!DOCTYPE html>
<html>
<script src="../resources/js-test-pre.js"></script>
<script id="library" type="x-shader/x-metal">
    #include <metal_stdlib>
    
    using namespace metal;
    
    struct Vertex
    {
        float4 position [[position]];
    };
    
    vertex Vertex vertex_main(uint vid [[vertex_id]])
    {
        Vertex v;
        switch (vid) {
        case 0:
            v.position = float4(-.75, -.75, 0, 1);
            break;
        case 1:
            v.position = float4(.75, -.75, 0, 1);
            break;
        case 2:
            v.position = float4(0, .75, 0, 1);
            break;
        default:
            v.position = float4(0, 0, 0, 1);
        }
        return v;
    }
    
    fragment float4 fragment_main(Vertex vertexIn [[stage_in]])
    {
        return float4(1.0, 0.0, 0.0, 1.0);
    }
</script>
<script>
'use strict';

if (window.testRunner)
    window.testRunner.dumpAsText();

let canvas;
let context;
let adapter;
let defaultDevice;

async function setUpContexts() {
    canvas = document.createElement("canvas");
    context = canvas.getContext("webgpu");
    if (!context)
        testFailed("Could not create WebGPU context!");

    shouldBeDefined(window.webgpu);
    
    adapter = await window.webgpu.requestAdapter({});
    if (!adapter) {
        testFailed("Could not create default WebGPUAdapter!")
        return;
    }

    defaultDevice = adapter.createDevice();
    if (!defaultDevice) {
        testFailed("Could not create WebGPUDevice!");
        return;
    }

    setUpShaders();
}

function setUpShaders() {
    let shaderDescriptor = {
        code: document.getElementById("library").text
    };
    let shaderModule = defaultDevice.createShaderModule(shaderDescriptor);
    if (!shaderModule)
        testFailed("Could not create WebGPUShaderModule!");
}

setUpContexts();

var successfullyParsed = true;
</script>
<script src="../resources/js-test-post.js"></script>
</html>
